+++ crates/miden-testing/tests/integration/scripts/faucet.rs
@@ -214,7 +214,7 @@ fn prove_faucet_contract_burn_fungible_asset_succeeds() -> anyhow::Result<()> {
     // Prove, serialize/deserialize and verify the transaction
     prove_and_verify_transaction(executed_transaction.clone())?;
 
-    assert_eq!(executed_transaction.account_delta().nonce_increment(), Felt::new(1));
+    assert_eq!(executed_transaction.account_delta().nonce_delta(), Felt::new(1));
     assert_eq!(executed_transaction.input_notes().get_note(0).id(), note.id());
     Ok(())
 }