+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Added procedure `was_procedure_called` to `miden::account` library module (#1521).
 - Implement serialization for `LexicographicWord` (#1524).
 - Make `Account:increment_nonce()` method public ([#1533](https://github.com/0xMiden/miden-base/pull/1533)).
+- Define the commitment to an empty account delta as `EMPTY_WORD` ([#1528](https://github.com/0xMiden/miden-base/pull/1528)).
 
 ## 0.9.5 (2025-06-20) - `miden-lib` crate only